Title: 거울에 레이저 쏘기 2, Time: 388ms, Memory: 24MB, Status: Passed - CodeTree

+++ 240206/거울에 레이저 쏘기 2/shoot-a-laser-in-the-mirror-2.java
@@ -27,9 +27,8 @@ public class Main {
             }
         }
         start = sc.nextInt();
-        float first = start / n;
+        float first = (float)start / n;
         int second = start % n;
-    
         if (first <= 1) {
             x = 0;
             d = 2;
